Ver 0.8.20
- More warnings squashed under Linux/MacOS.

+++ Contrib/at67/expression.cpp
@@ -213,7 +213,7 @@ namespace Expression
         if(varName.size() == 0)  return false;
         if(!isalpha((unsigned char)varName[0])) return false;
 
-        for(int i=1; i<varName.size()-1; i++)
+        for(size_t i=1; i<varName.size()-1; i++)
         {
             if(!isalnum((unsigned char)varName[i])) return false;
         }
@@ -230,7 +230,7 @@ namespace Expression
         if(str.size() < 2) return false;
         if(str[0] == '"'  &&  str.back() == '"')
         {
-            for(int i=1; i<str.size()-1; i++)
+            for(size_t i=1; i<str.size()-1; i++)
             {
                 if(str[i] == '"') return false;
             }
@@ -379,7 +379,7 @@ namespace Expression
         int spaceCount = 0;
         bool inString = false;
 
-        for(int i=0; i<input.size(); i++)
+        for(size_t i=0; i<input.size(); i++)
         {
             if(input[i] == '\"') inString = !inString;
 
@@ -410,7 +410,7 @@ namespace Expression
         bool inString = false;
         bool inComment = false;
 
-        for(int i=0; i<input.size(); i++)
+        for(size_t i=0; i<input.size(); i++)
         {
             // Check for string
             if(!inComment  &&  input[i] == '\"') inString = !inString;
@@ -418,7 +418,7 @@ namespace Expression
             // Check for comment, ' and REM
             if(!inString)
             {
-                if((input[i] == '\'')  ||  (i <= input.size() - 3  &&  toupper((unsigned char)input[i]) == 'R'  &&  toupper((unsigned char)input[i+1]) == 'E'  &&  toupper((unsigned char)input[i+2]) == 'M'))
+                if((input[i] == '\'')  ||  (i <= int(input.size()) - 3  &&  toupper((unsigned char)input[i]) == 'R'  &&  toupper((unsigned char)input[i+1]) == 'E'  &&  toupper((unsigned char)input[i+2]) == 'M'))
                 {
                     inComment = true;
                 }
@@ -444,7 +444,7 @@ namespace Expression
 
     void padString(std::string &input, int num, char pad)
     {
-        if(num > input.size()) input.insert(0, num - input.size(), pad);
+        if(num > int(input.size())) input.insert(0, num - int(input.size()), pad);
     }
 
     void addString(std::string &input, int num, char add)
@@ -455,7 +455,7 @@ namespace Expression
     int tabbedStringLength(const std::string& input, int tabSize)
     {
         int length = 0, newLine = 0;
-        for(int i=0; i<input.size(); i++)
+        for(int i=0; i<int(input.size()); i++)
         {
             switch(input[i])
             {
@@ -607,7 +607,7 @@ namespace Expression
         // Hex
         if(token.size() >= 2  &&  token.c_str()[0] == '$')
         {
-            for(int i=1; i<token.size(); i++) success &= _hexaDecimalChars[token.c_str()[i]];
+            for(size_t i=1; i<token.size(); i++) success &= _hexaDecimalChars[uint8_t(token.c_str()[i])];
             if(success)
             {
                 result = strtol(&token.c_str()[1], NULL, 16);
@@ -617,7 +617,7 @@ namespace Expression
         // Hex
         else if(token.size() >= 3  &&  ((token.c_str()[0] == '0'  &&  token.c_str()[1] == 'X')  ||  (token.c_str()[0] == '&'  &&  token.c_str()[1] == 'H')))
         {
-            for(int i=2; i<token.size(); i++) success &= _hexaDecimalChars[token.c_str()[i]];
+            for(size_t i=2; i<token.size(); i++) success &= _hexaDecimalChars[uint8_t(token.c_str()[i])];
             if(success)
             {
                 result = strtol(&token.c_str()[2], NULL, 16);
@@ -627,7 +627,7 @@ namespace Expression
         // Octal
         else if(token.size() >= 3  &&  ((token.c_str()[0] == '0'  &&  (token.c_str()[1] == 'O' || token.c_str()[1] == 'Q'))  ||  (token.c_str()[0] == '&'  &&  token.c_str()[1] == 'O')))
         {
-            for(int i=2; i<token.size(); i++) success &= _octalChars[token.c_str()[i]];
+            for(size_t i=2; i<token.size(); i++) success &= _octalChars[uint8_t(token.c_str()[i])];
             if(success)
             {
                 result = strtol(&token.c_str()[2], NULL, 8);
@@ -637,7 +637,7 @@ namespace Expression
         // Binary
         else if(token.size() >= 3  &&  ((token.c_str()[0] == '0'  &&  token.c_str()[1] == 'B')  ||  (token.c_str()[0] == '&'  &&  token.c_str()[1] == 'B')))
         {
-            for(int i=2; i<token.size(); i++) success &= _binaryChars[token.c_str()[i]];
+            for(size_t i=2; i<token.size(); i++) success &= _binaryChars[uint8_t(token.c_str()[i])];
             if(success)
             {
                 result = strtol(&token.c_str()[2], NULL, 2);
@@ -647,7 +647,7 @@ namespace Expression
         // Decimal
         else
         {
-            for(int i=0; i<token.size(); i++) success &= _decimalChars[token.c_str()[i]];
+            for(size_t i=0; i<token.size(); i++) success &= _decimalChars[uint8_t(token.c_str()[i])];
             if(success)
             {
                 result = strtol(&token.c_str()[0], NULL, 10);
@@ -842,7 +842,7 @@ namespace Expression
         DelimiterState delimiterState = WhiteSpace;
         std::vector<std::string> tokens;
 
-        for(int i=0; i<=line.size(); i++)
+        for(size_t i=0; i<=line.size(); i++)
         {
             // End of line is a delimiter for white space
             if(i == line.size())
@@ -930,7 +930,7 @@ namespace Expression
         DelimiterState delimiterState = WhiteSpace;
         std::vector<std::string> tokens;
 
-        for(int i=0; i<=line.size(); i++)
+        for(size_t i=0; i<=line.size(); i++)
         {
             // End of line is a delimiter for white space
             if(i == line.size())
